Working on the UI. The layout is similar to VS code or Obsidian
clearly not as complex. Currently all created quizzes in the DB
will populate as a list of files within the InfoDrawer component.
This is basic no folder creation or drag/drop has been added yet
In the main area the selected quiz details are being displayed as
JSON. This will be refactored as more functionality is added.