Refactoring.
Removed the original side-toggler as the folder icon seemed more
intuitive. Created a VirtualComponent for the virtual filesystem
this wrapper grabs the info needed for a right-click menu and
controls the logic for when the menu should be rendered. Options
like rename/delete etc. The functionality has not implemented yet
in the drop-down component but it renders accordingly