Addressed some styling issues with the info drawer and quiz topics
refactored the InputModal,this is as work in progress. Added
chevrons next to the folders to make it easier to tell if a folder
is open or closed. While we have open and closed folders they look
similar. Refactored the AddFolder in the toolbar to use the
AddFolderToRoot component which sets vars for the InputModal.
In the context menu for folders, the option for adding a folder
has been implemented, using the AddFolderToFolder which sets vars
for the input modal, powered by a signal.