Automating creating a quiz from and existing quiz.
Text-file copies of quizzes from UofA or TestOut can be converted
to the json format needed to create a quiz by json from the rest
endpoint. Data is extracted manually from the provided text and
OpenAIs GPT-4 is used to fill in the gaps. The AI generated
content needs to be verified before creating a quiz but the UI
is ready to implemenent the create quiz feature.