The add quiz functionality from the side bar has been implemented
Quizzes can be created by uploading a txt file of the completed
quiz. If a quiz has more than 10 questions it will be broken into
multiple quizzes that have a max of 10 questions each. A folder is
automatically created using the quiz name and the quizzes are added
to it. The created items are added to the virtual filesystem and the
add quiz modal is closed